Cleaning and reformatting
