+++ cmd/aws/main.go
@@ -1,16 +1,18 @@
 package main
 
 import (
+	"bytes"
+	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"image/jpeg"
 	"log"
 	"net/http"
 
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
 
-	"github.com/josemyduarte/printer"
-	"github.com/josemyduarte/printer/internal/aws"
+	"github.com/josemyduarte/printer/internal/printer"
 )
 
 var (
@@ -21,6 +23,8 @@ var (
 	bgFileName = "/tmp/00-instagram-background.png"
 )
 
+// It seems that in AWS lambdas you don't have access to the asset folder.
+// Downloading the assets on init to make sure we have it available on runtime.
 func init() {
 	if err := downloadFile(fontURL, fontFileName); err != nil {
 		panic(fmt.Errorf("couldn't download font: %w", err))
@@ -46,12 +50,38 @@ func handler(request events.APIGatewayProxyRequest) (events.APIGatewayProxyRespo
 		}, err
 	}
 
-	return aws.Serve(printer.Request{
+	img, err := printer.TextOnImg(printer.Request{
 		BgImgPath: bgFileName,
 		FontPath:  fontFileName,
 		FontSize:  60,
 		Text:      req.Text,
 	})
+	if err != nil {
+		log.Println(err)
+		return events.APIGatewayProxyResponse{
+			StatusCode: http.StatusInternalServerError,
+			Body:       err.Error(),
+		}, err
+	}
+
+	imgBuf := new(bytes.Buffer)
+	if jpeg.Encode(imgBuf, img, nil) != nil {
+		log.Println(err)
+		return events.APIGatewayProxyResponse{
+			StatusCode: http.StatusInternalServerError,
+			Body:       err.Error(),
+		}, err
+	}
+
+	return events.APIGatewayProxyResponse{
+		StatusCode:      http.StatusOK,
+		Body:            base64.StdEncoding.EncodeToString(imgBuf.Bytes()),
+		IsBase64Encoded: true,
+		Headers: map[string]string{
+			"Content-Type": "image/png",
+		},
+	}, nil
+
 }
 
 func main() {
